+++ common/bandwidth-controller/src/acquire/state.rs
@@ -1,8 +1,8 @@
-// Copyright 2022 - Nym Technologies SA <contact@nymtech.net>
+// Copyright 2022-2023 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
 use nym_coconut_interface::Parameters;
-use nym_credentials::coconut::bandwidth::BandwidthVoucher;
+use nym_credentials::coconut::bandwidth::{BandwidthVoucher, TOTAL_ATTRIBUTES};
 
 use nym_crypto::asymmetric::{encryption, identity};
 
@@ -29,7 +29,16 @@ impl From<encryption::KeyPair> for KeyPair {
     }
 }
 
-pub(crate) struct State {
+pub struct State {
     pub voucher: BandwidthVoucher,
     pub params: Parameters,
+}
+
+impl State {
+    pub fn new(voucher: BandwidthVoucher) -> Self {
+        State {
+            voucher,
+            params: Parameters::new(TOTAL_ATTRIBUTES).unwrap(),
+        }
+    }
 }