+++ common/credential-storage/src/error.rs
@@ -5,9 +5,11 @@ use thiserror::Error;
 
 #[derive(Error, Debug)]
 pub enum StorageError {
+    #[cfg(not(target_arch = "wasm32"))]
     #[error("Database experienced an internal error - {0}")]
     InternalDatabaseError(#[from] sqlx::Error),
 
+    #[cfg(not(target_arch = "wasm32"))]
     #[error("Failed to perform database migration - {0}")]
     MigrationError(#[from] sqlx::migrate::MigrateError),