+++ CHANGELOG.md
@@ -6,10 +6,12 @@ Post 1.0.0 release, the changelog format is based on [Keep a Changelog](https://
 
 - nym-network-statistics properly handles signals ([#3209])
 - add socks5 support for Rust SDK ([#3226], [#3255])
+- add coconut bandwidth credential support for Rust SDK ([#3273]) 
 
 [#3209]: https://github.com/nymtech/nym/issues/3209
 [#3226]: https://github.com/nymtech/nym/pull/3226
 [#3255]: https://github.com/nymtech/nym/pull/3255
+[#3273]: https://github.com/nymtech/nym/pull/3273
 
 ## [v1.1.13] (2023-03-15)